add a connection pool for faster and simoutaneous processing
